+++ src/modules/graphics/opengl.c
@@ -2612,14 +2612,14 @@ Shader* lovrShaderCreateGraphics(const char* vertexSource, int vertexSourceLengt
   vertexSource = vertexSource == NULL ? lovrUnlitVertexShader : vertexSource;
   const char* vertexSources[] = { version, singlepass[0], flagSource ? flagSource : "", lovrShaderVertexPrefix, vertexSource, lovrShaderVertexSuffix };
   int vertexSourceLengths[] = { -1, -1, -1, -1, vertexSourceLength, -1 };
-  size_t vertexSourceCount = sizeof(vertexSources) / sizeof(vertexSources[0]);
+  int vertexSourceCount = sizeof(vertexSources) / sizeof(vertexSources[0]);
   GLuint vertexShader = compileShader(GL_VERTEX_SHADER, vertexSources, vertexSourceLengths, vertexSourceCount);
 
   // Fragment
   fragmentSource = fragmentSource == NULL ? lovrUnlitFragmentShader : fragmentSource;
   const char* fragmentSources[] = { version, singlepass[1], flagSource ? flagSource : "", lovrShaderFragmentPrefix, fragmentSource, lovrShaderFragmentSuffix };
   int fragmentSourceLengths[] = { -1, -1, -1, -1, fragmentSourceLength, -1 };
-  size_t fragmentSourceCount = sizeof(fragmentSources) / sizeof(fragmentSources[0]);
+  int fragmentSourceCount = sizeof(fragmentSources) / sizeof(fragmentSources[0]);
   GLuint fragmentShader = compileShader(GL_FRAGMENT_SHADER, fragmentSources, fragmentSourceLengths, fragmentSourceCount);
 
   free(flagSource);
@@ -2695,7 +2695,7 @@ Shader* lovrShaderCreateCompute(const char* source, int length, ShaderFlag* flag
   char* flagSource = lovrShaderGetFlagCode(flags, flagCount);
   const char* sources[] = { lovrShaderComputePrefix, flagSource ? flagSource : "", source, lovrShaderComputeSuffix };
   int lengths[] = { -1, -1, length, -1 };
-  size_t count = sizeof(sources) / sizeof(sources[0]);
+  int count = sizeof(sources) / sizeof(sources[0]);
   GLuint computeShader = compileShader(GL_COMPUTE_SHADER, sources, lengths, count);
   free(flagSource);
   GLuint program = glCreateProgram();
@@ -2842,7 +2842,7 @@ size_t lovrShaderComputeUniformLayout(arr_uniform_t* uniforms) {
 ShaderBlock* lovrShaderBlockCreate(BlockType type, Buffer* buffer, arr_uniform_t* uniforms) {
   ShaderBlock* block = lovrAlloc(ShaderBlock);
   arr_init(&block->uniforms);
-  map_init(&block->uniformMap, uniforms->length);
+  map_init(&block->uniformMap, (uint32_t) uniforms->length);
 
   arr_append(&block->uniforms, uniforms->data, uniforms->length);
