+++ src/modules/thread/thread.h
@@ -18,7 +18,7 @@ typedef struct Thread {
   mtx_t lock;
   Blob* body;
   Variant arguments[MAX_THREAD_ARGUMENTS];
-  size_t argumentCount;
+  uint32_t argumentCount;
   int (*runner)(void*);
   char* error;
   bool running;
@@ -32,7 +32,7 @@ void lovrThreadRemoveChannel(uint64_t hash);
 Thread* lovrThreadInit(Thread* thread, int (*runner)(void*), Blob* body);
 #define lovrThreadCreate(...) lovrThreadInit(lovrAlloc(Thread), __VA_ARGS__)
 void lovrThreadDestroy(void* ref);
-void lovrThreadStart(Thread* thread, Variant* arguments, size_t argumentCount);
+void lovrThreadStart(Thread* thread, Variant* arguments, uint32_t argumentCount);
 void lovrThreadWait(Thread* thread);
 const char* lovrThreadGetError(Thread* thread);
 bool lovrThreadIsRunning(Thread* thread);
